+++ Runtime/ConvertFlags.cs
@@ -11,6 +11,10 @@ namespace UnderLogic.Serialization.Toml
         ForceArray = 0x4,
         ForceList = 0x8,
         ForceInline = 0x10,
-        ForceExpand = 0x20
+        ForceExpand = 0x20,
+        HexNumberLowerCase = 0x40,
+        HexNumberUpperCase = 0x80,
+        OctalNumber = 0x100,
+        BinaryNumber = 0x200,
     }
 }