+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.4] - 2023-01-02
 
-### Fixed
+### Added
 
-
+- Serialization support for `PositiveInfinity`, `NegativeInfinity` and `NaN` values
+- Deserialization support for `PositiveInfinity`, `NegativeInfinity` and `NaN` values
 
 ## [0.4.3] - 2022-12-31